changes to support large objects
git-svn-id: https://atmos-c.googlecode.com/svn/trunk@23 cb61ccbc-2dfe-265f-a3f4-a677fceac7d0

+++ transport.h
@@ -11,10 +11,10 @@ typedef struct credentialsval {
 
 typedef struct PD {
     char *data;
-    size_t body_size;
-    size_t bytes_remaining;
-    size_t offset;
-    size_t bytes_written;
+    unsigned long long body_size;
+    unsigned long long bytes_remaining;
+    unsigned long long offset;
+    unsigned long long bytes_written;
 }postdata;
 
 typedef struct hdrval {
